Splitting a single file. View binding.
