+++ app/src/main/java/com/gb/stopwatch/ElapsedTimeCalculator.kt
@@ -1,0 +1,16 @@
+package com.gb.stopwatch
+
+class ElapsedTimeCalculator(
+    private val timestampProvider: TimestampProvider,
+) {
+
+    fun calculate(state: StopwatchState.Running): Long {
+        val currentTimestamp = timestampProvider.getMilliseconds()
+        val timePassedSinceStart = if (currentTimestamp > state.startTime) {
+            currentTimestamp - state.startTime
+        } else {
+            0
+        }
+        return timePassedSinceStart + state.elapsedTime
+    }
+}
