+++ app/src/main/java/com/gb/stopwatch/StopwatchListOrchestrator.kt
@@ -1,0 +1,49 @@
+package com.gb.stopwatch
+
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class StopwatchListOrchestrator(
+    private val stopwatchStateHolder: StopwatchStateHolder,
+    private val scope: CoroutineScope,
+) {
+
+    private var job: Job? = null
+    private val mutableTicker = MutableStateFlow("")
+    val ticker: StateFlow<String> = mutableTicker
+
+    fun start() {
+        if (job == null) startJob()
+        stopwatchStateHolder.start()
+    }
+
+    private fun startJob() {
+        scope.launch {
+            while (isActive) {
+                mutableTicker.value = stopwatchStateHolder.getStringTimeRepresentation()
+                delay(20)
+            }
+        }
+    }
+
+    fun pause() {
+        stopwatchStateHolder.pause()
+        stopJob()
+    }
+
+    fun stop() {
+        stopwatchStateHolder.stop()
+        stopJob()
+        clearValue()
+    }
+
+    private fun stopJob() {
+        scope.coroutineContext.cancelChildren()
+        job = null
+    }
+
+    private fun clearValue() {
+        mutableTicker.value = "00:00:000"
+    }
+}
